api key required

+++ app/api/item/dynamic/[itemId]/route.ts
@@ -1,5 +1,7 @@
 import { NextResponse } from "next/server";
 import { connectToMongo } from "../../../middleware/connectToMongo";
+import keyAuth from "@/app/api/middleware/keyAuth";
+import { headers } from "next/headers";
 const mongoose = require("mongoose");
 const Item = require("../../../../../models/itemsModel");
 
@@ -48,31 +50,40 @@ export async function GET(request: Request) {
 
 // delete an item
 export async function DELETE(request: Request) {
+  const headersList = headers();
+  const apiKey = headersList.get("apiKey");
   try {
-    connectToMongo();
-    const url = request.url;
-    const splitUrl = url.split("/");
-    const id = splitUrl[splitUrl.length - 1];
-    if (!mongoose.Types.ObjectId.isValid(id)) {
-      return NextResponse.json(
-        { error: "No such item" },
-        {
-          status: 404,
-        }
-      );
-    }
+    if (keyAuth(apiKey) === false) {
+      return NextResponse.json({
+        error: "api key is missing or incorrect",
+        status: 404,
+      });
+    } else {
+      connectToMongo();
+      const url = request.url;
+      const splitUrl = url.split("/");
+      const id = splitUrl[splitUrl.length - 1];
+      if (!mongoose.Types.ObjectId.isValid(id)) {
+        return NextResponse.json(
+          { error: "No such item" },
+          {
+            status: 404,
+          }
+        );
+      }
 
-    const item = await Item.findOneAndDelete({ _id: id });
+      const item = await Item.findOneAndDelete({ _id: id });
 
-    if (!item) {
-      return NextResponse.json(
-        { error: "No such item" },
-        {
-          status: 404,
-        }
-      );
+      if (!item) {
+        return NextResponse.json(
+          { error: "No such item" },
+          {
+            status: 404,
+          }
+        );
+      }
+      return NextResponse.json(item);
     }
-    return NextResponse.json(item);
   } catch (error: unknown) {
     if (error instanceof Error)
       return (
@@ -86,37 +97,46 @@ export async function DELETE(request: Request) {
 
 // update an item
 export async function PATCH(request: Request) {
+  const headersList = headers();
+  const apiKey = headersList.get("apiKey");
   try {
-    connectToMongo();
-    const url = request.url;
-    const splitUrl = url.split("/");
-    const id = splitUrl[splitUrl.length - 1];
-    const req: Item = await request.json();
-    if (!mongoose.Types.ObjectId.isValid(id)) {
-      return NextResponse.json(
-        { error: "No such item" },
-        {
-          status: 404,
-        }
-      );
-    }
-
-    const item = await Item.findOneAndUpdate(
-      { _id: id },
-      {
-        ...req,
+    if (keyAuth(apiKey) === false) {
+      return NextResponse.json({
+        error: "api key is missing or incorrect",
+        status: 404,
+      });
+    } else {
+      connectToMongo();
+      const url = request.url;
+      const splitUrl = url.split("/");
+      const id = splitUrl[splitUrl.length - 1];
+      const req: Item = await request.json();
+      if (!mongoose.Types.ObjectId.isValid(id)) {
+        return NextResponse.json(
+          { error: "No such item" },
+          {
+            status: 404,
+          }
+        );
       }
-    );
 
-    if (!item) {
-      return NextResponse.json(
-        { error: "No such item" },
+      const item = await Item.findOneAndUpdate(
+        { _id: id },
         {
-          status: 404,
+          ...req,
         }
       );
+
+      if (!item) {
+        return NextResponse.json(
+          { error: "No such item" },
+          {
+            status: 404,
+          }
+        );
+      }
+      return NextResponse.json(item);
     }
-    return NextResponse.json(item);
   } catch (error: unknown) {
     if (error instanceof Error)
       return NextResponse.json(
