+++ app/api/user/dynamic/[userId]/route.ts
@@ -1,6 +1,8 @@
 import { NextResponse } from "next/server";
 import { useRouter } from "next/router";
 import { connectToMongo } from "@/app/api/middleware/connectToMongo";
+import { headers } from "next/headers";
+import keyAuth from "@/app/api/middleware/keyAuth";
 const mongoose = require("mongoose");
 const User = require("../../../../../models/usersModel");
 
@@ -53,31 +55,40 @@ export async function GET(request: Request) {
 
 //delete a user
 export async function DELETE(request: Request) {
+  const headersList = headers();
+  const apiKey = headersList.get("apiKey");
   try {
-    connectToMongo();
-    const url = request.url;
-    const splitUrl = url.split("/");
-    const id = splitUrl[splitUrl.length - 1];
-    if (!mongoose.Types.ObjectId.isValid(id)) {
-      return NextResponse.json(
-        { error: "No such user" },
-        {
-          status: 404,
-        }
-      );
-    }
+    if (keyAuth(apiKey) === false) {
+      return NextResponse.json({
+        error: "api key is missing or incorrect",
+        status: 404,
+      });
+    } else {
+      connectToMongo();
+      const url = request.url;
+      const splitUrl = url.split("/");
+      const id = splitUrl[splitUrl.length - 1];
+      if (!mongoose.Types.ObjectId.isValid(id)) {
+        return NextResponse.json(
+          { error: "No such user" },
+          {
+            status: 404,
+          }
+        );
+      }
 
-    const user = await User.findOneAndDelete({ _id: id });
+      const user = await User.findOneAndDelete({ _id: id });
 
-    if (!user) {
-      return NextResponse.json(
-        { error: "No such user" },
-        {
-          status: 404,
-        }
-      );
+      if (!user) {
+        return NextResponse.json(
+          { error: "No such user" },
+          {
+            status: 404,
+          }
+        );
+      }
+      return NextResponse.json(user);
     }
-    return NextResponse.json(user);
   } catch (error: unknown) {
     if (error instanceof Error)
       return NextResponse.json(
@@ -92,37 +103,46 @@ export async function DELETE(request: Request) {
 //update a user
 // needs to add token hashing
 export async function PATCH(request: Request) {
+  const headersList = headers();
+  const apiKey = headersList.get("apiKey");
   try {
-    connectToMongo();
-    const url = request.url;
-    const splitUrl = url.split("/");
-    const id = splitUrl[splitUrl.length - 1];
-    const req: Item = await request.json();
-    if (!mongoose.Types.ObjectId.isValid(id)) {
-      return NextResponse.json(
-        { error: "No such user" },
-        {
-          status: 404,
-        }
-      );
-    }
-
-    const user = await User.findOneAndUpdate(
-      { _id: id },
-      {
-        ...req,
+    if (keyAuth(apiKey) === false) {
+      return NextResponse.json({
+        error: "api key is missing or incorrect",
+        status: 404,
+      });
+    } else {
+      connectToMongo();
+      const url = request.url;
+      const splitUrl = url.split("/");
+      const id = splitUrl[splitUrl.length - 1];
+      const req: Item = await request.json();
+      if (!mongoose.Types.ObjectId.isValid(id)) {
+        return NextResponse.json(
+          { error: "No such user" },
+          {
+            status: 404,
+          }
+        );
       }
-    );
 
-    if (!user) {
-      return NextResponse.json(
-        { error: "No such user" },
+      const user = await User.findOneAndUpdate(
+        { _id: id },
         {
-          status: 404,
+          ...req,
         }
       );
+
+      if (!user) {
+        return NextResponse.json(
+          { error: "No such user" },
+          {
+            status: 404,
+          }
+        );
+      }
+      return NextResponse.json(user);
     }
-    return NextResponse.json(user);
   } catch (error: unknown) {
     if (error instanceof Error)
       return NextResponse.json(
